+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - [@HiromuHota][HiromuHota]: Fix a bug that an html file is not created at a given path.
   ([#64](https://github.com/HazyResearch/pdftotree/pull/64))
-
+- [@HiromuHota][HiromuHota]: Switch the output format from "HTML-like" to hOCR.
+  ([#62](https://github.com/HazyResearch/pdftotree/pull/62))
 
 ## 0.4.1 - 2020-09-21